+++ libafl/src/events/launcher.rs
@@ -57,6 +57,7 @@ use crate::{
     events::{
         llmp::{LlmpRestartingEventManager, LlmpShouldSaveState, ManagerKind, RestartingMgr},
         EventConfig,
+        centralized::CentralizedEventManagerBuilder,
     },
     monitors::Monitor,
     state::{HasExecutions, State},
@@ -696,19 +697,23 @@ where
                         let builder = builder.time_ref(self.time_obs.handle());
                         let (state, mgr) = builder.build().launch()?;
 
+                        let mut centralized_builder = CentralizedEventManagerBuilder::new();
+
+                        if index == 1 {
+                            centralized_builder = centralized_builder.is_main(true);
+                        }
+
                         #[cfg(not(feature = "adaptive_serialization"))]
-                        let c_mgr = CentralizedEventManager::on_port(
+                        let c_mgr = centralized_builder.build_on_port(
                             mgr,
                             self.shmem_provider.clone(),
                             self.centralized_broker_port,
-                            index == 1,
                         )?;
                         #[cfg(feature = "adaptive_serialization")]
-                        let c_mgr = CentralizedEventManager::on_port(
+                        let c_mgr = centralized_builder.build_on_port(
                             mgr,
                             self.shmem_provider.clone(),
                             self.centralized_broker_port,
-                            index == 1,
                             self.time_obs,
                         )?;
